added internet perms + launch thing
corrects the launcher to launch login first. it was already doing it, but  main activity was also a launcher- which just removed for future sake so im not scratching my head. also, just as a note, the main activity page has 'hello, user!" i cant find where this is, but we can set it to be the email instead :D